renamed src folder to lib
