+++ lib/auth.ts
