+++ lib/buffer.ts
