+++ lib/constant/errors.ts
