+++ lib/constant/mysql_types.ts
