+++ lib/deferred.ts
