+++ lib/packets/builders/auth.ts
