+++ lib/packets/builders/query.ts
