+++ lib/packets/parsers/authswitch.ts
