+++ lib/packets/parsers/err.ts
