+++ lib/packets/parsers/result.ts
