+++ lib/pool.ts
