+++ lib/util.ts
